js: Add support for ephemeral subscribe
