+++ nats/aio/client.py
@@ -30,6 +30,7 @@ from dataclasses import dataclass, field
 from nats.errors import *
 from nats.aio.errors import *
 from nats.nuid import NUID
+from nats.aio.subscription import *
 from nats.protocol.parser import *
 from nats.protocol import command as prot_command
 from nats.js import JetStream, JetStreamContext, JetStreamManager
@@ -69,10 +70,6 @@ DEFAULT_CONNECT_TIMEOUT = 2  # in seconds
 DEFAULT_DRAIN_TIMEOUT = 30  # in seconds
 MAX_CONTROL_LINE_SIZE = 1024
 
-# Default Pending Limits of Subscriptions
-DEFAULT_SUB_PENDING_MSGS_LIMIT = 512 * 1024
-DEFAULT_SUB_PENDING_BYTES_LIMIT = 128 * 1024 * 1024
-
 NATS_HDR_LINE = bytearray(b'NATS/1.0\r\n')
 NATS_HDR_LINE_SIZE = len(NATS_HDR_LINE)
 NO_RESPONDERS_STATUS = "503"
@@ -80,268 +77,6 @@ STATUS_MSG_LEN = 3  # e.g. 20x, 40x, 50x
 CTRL_LEN = len(_CRLF_)
 STATUS_HDR = "Status"
 DESC_HDR = "Description"
-
-
-class Subscription:
-    """
-    A Subscription represents interest in a particular subject.
-
-    A Subscription should not be constructed directly, rather
-    `connection.subscribe()` should be used to get a subscription.
-
-    ::
-
-        nc = nats.connect()
-
-        # Async Subscription
-        async def cb(msg):
-          print('Received', msg)
-        await nc.subscribe('foo', cb=cb)
-
-        # Sync Subscription
-        sub = nc.subscribe('foo')
-        msg = await sub.next_msg()
-        print('Received', msg)
-    
-    """
-    def __init__(
-        self,
-        conn: 'Client',
-        id: int = 0,
-        subject: str = '',
-        queue: str = '',
-        cb: Optional[Callable[['Msg'], None]] = None,
-        future: Optional[asyncio.Future] = None,
-        max_msgs: int = 0,
-        pending_msgs_limit: int = DEFAULT_SUB_PENDING_MSGS_LIMIT,
-        pending_bytes_limit: int = DEFAULT_SUB_PENDING_BYTES_LIMIT,
-    ):
-        self._conn = conn
-        self._id = id
-        self._subject = subject
-        self._queue = queue
-        self._max_msgs = max_msgs
-        self._received = 0
-        self._cb = cb
-        self._future = future
-        self._closed = False
-
-        # Per subscription message processor.
-        self._pending_msgs_limit = pending_msgs_limit
-        self._pending_bytes_limit = pending_bytes_limit
-        self._pending_queue = asyncio.Queue(maxsize=pending_msgs_limit)
-        self._pending_size = 0
-        self._wait_for_msgs_task = None
-        self._message_iterator = None
-
-    @property
-    def subject(self) -> str:
-        """
-        Returns the subject of the `Subscription`.
-        """
-        return self._subject
-
-    @property
-    def queue(self) -> str:
-        """
-        Returns the queue name of the `Subscription` if part of a queue group.
-        """
-        return self._queue
-
-    @property
-    def messages(self) -> AsyncIterator['Msg']:
-        """
-        Retrieves an async iterator for the messages from the subscription.
-
-        This is only available if a callback isn't provided when creating a
-        subscription.
-        """
-        if not self._message_iterator:
-            raise NatsError(
-                "cannot iterate over messages with a non iteration subscription type"
-            )
-
-        return self._message_iterator
-
-    @property
-    def pending_msgs(self) -> int:
-        return self._pending_queue.qsize()
-
-    async def next_msg(self, timeout: float = 1.0):
-        """
-        :params timeout: Time to wait for next message before
-        :raises nats.errors.TimeoutError:
-
-        next_msg can be used to retrieve the next message
-        from a stream of messages using await syntax, this
-        only works when not passing a callback on `subscribe`::
-
-            sub = await nc.subscribe('hello')
-            msg = await sub.next_msg(timeout=1)
-
-        """
-        future = asyncio.Future()
-
-        async def _next_msg():
-            msg = await self._pending_queue.get()
-            future.set_result(msg)
-
-        task = asyncio.get_running_loop().create_task(_next_msg())
-        try:
-            msg = await asyncio.wait_for(future, timeout)
-            return msg
-        except asyncio.TimeoutError:
-            future.cancel()
-            task.cancel()
-            raise ErrTimeout
-
-    def _start(self, error_cb):
-        """
-        Creates the resources for the subscription to start processing messages.
-        """
-        if self._cb:
-            if not asyncio.iscoroutinefunction(self._cb) and \
-                not (hasattr(self._cb, "func") and asyncio.iscoroutinefunction(self._cb.func)):
-                raise NatsError("nats: must use coroutine for subscriptions")
-
-            self._wait_for_msgs_task = asyncio.get_running_loop().create_task(
-                self._wait_for_msgs(error_cb)
-            )
-
-        elif self._future:
-            # Used to handle the single response from a request.
-            pass
-        else:
-            self._message_iterator = _SubscriptionMessageIterator(
-                self._pending_queue
-            )
-
-    async def drain(self):
-        """
-        Removes interest in a subject, but will process remaining messages.
-        """
-        if self._conn.is_closed:
-            raise ConnectionClosedError
-        if self._conn.is_draining:
-            raise ConnectionDrainingError
-        if self._closed:
-            raise BadSubscriptionError
-
-        try:
-            # Announce server that no longer want to receive more
-            # messages in this sub and just process the ones remaining.
-            await self._conn._send_unsubscribe(self._id)
-
-            # Roundtrip to ensure that the server has sent all messages.
-            await self._conn.flush()
-
-            if self._pending_queue:
-                # Wait until no more messages are left,
-                # then cancel the subscription task.
-                await self._pending_queue.join()
-
-            # stop waiting for messages
-            self._stop_processing()
-
-            # Subscription is done and won't be receiving further
-            # messages so can throw it away now.
-            self._conn._remove_sub(self._id)
-        except asyncio.CancelledError:
-            # In case draining of a connection times out then
-            # the sub per task will be canceled as well.
-            pass
-        finally:
-            self._closed = True
-
-    async def unsubscribe(self, limit: int = 0):
-        """
-        :param limit: Max number of messages to receive before unsubscribing.
-
-        Removes interest in a subject, remaining messages will be discarded.
-
-        If `limit` is greater than zero, interest is not immediately removed,
-        rather, interest will be automatically removed after `limit` messages
-        are received.
-        """
-        if self._conn.is_closed:
-            raise ConnectionClosedError
-        if self._conn.is_draining:
-            raise ConnectionDrainingError
-        if self._closed:
-            raise BadSubscriptionError
-
-        self._max_msgs = limit
-        if limit == 0 or self._received >= limit:
-            self._stop_processing()
-            self._conn._remove_sub(self._id)
-
-        if not self._conn.is_reconnecting:
-            await self._conn._send_unsubscribe(self._id, limit=limit)
-        self._closed = True
-
-    def _stop_processing(self):
-        """
-        Stops the subscription from processing new messages.
-        """
-        if self._wait_for_msgs_task and not self._wait_for_msgs_task.done():
-            self._wait_for_msgs_task.cancel()
-        if self._message_iterator:
-            self._message_iterator._cancel()
-
-    async def _wait_for_msgs(self, error_cb):
-        """
-        A coroutine to read and process messages if a callback is provided.
-
-        Should be called as a task.
-        """
-        while True:
-            try:
-                msg = await self._pending_queue.get()
-                self._pending_size -= len(msg.data)
-
-                try:
-                    # Invoke depending of type of handler.
-                    await self._cb(msg)
-                except asyncio.CancelledError:
-                    # In case the coroutine handler gets cancelled
-                    # then stop task loop and return.
-                    break
-                except Exception as e:
-                    # All errors from calling a handler
-                    # are async errors.
-                    if error_cb:
-                        await error_cb(e)
-                finally:
-                    # indicate the message finished processing so drain can continue
-                    self._pending_queue.task_done()
-
-            except asyncio.CancelledError:
-                break
-
-
-class _SubscriptionMessageIterator:
-    def __init__(self, queue):
-        self._queue = queue
-        self._unsubscribed_future = asyncio.Future()
-
-    def _cancel(self):
-        if not self._unsubscribed_future.done():
-            self._unsubscribed_future.set_result(True)
-
-    def __aiter__(self):
-        return self
-
-    async def __anext__(self):
-        get_task = asyncio.get_running_loop().create_task(self._queue.get())
-        finished, _ = await asyncio.wait([get_task, self._unsubscribed_future],
-                                         return_when=asyncio.FIRST_COMPLETED)
-        if get_task in finished:
-            self._queue.task_done()
-            return get_task.result()
-        elif self._unsubscribed_future.done():
-            get_task.cancel()
-
-        raise StopAsyncIteration
 
 
 class Msg:
@@ -961,16 +696,20 @@ class Client:
         if self.is_connecting or self.is_reconnecting:
             raise ErrConnectionReconnecting
 
-        # Start draining the subscriptions
-        self._status = Client.DRAINING_SUBS
-
         drain_tasks = []
         for sub in self._subs.values():
-            coro = sub.drain()
+            coro = sub._drain()
             task = asyncio.get_running_loop().create_task(coro)
             drain_tasks.append(task)
 
         drain_is_done = asyncio.gather(*drain_tasks)
+
+        # Start draining the subscriptions.
+        # Relinquish CPU to allow drain tasks to start in the background,
+        # before setting state to draining.
+        await asyncio.sleep(0)
+        self._status = Client.DRAINING_SUBS
+
         try:
             await asyncio.wait_for(
                 drain_is_done, self.options["drain_timeout"]
@@ -1213,6 +952,9 @@ class Client:
 
         try:
             msg = await asyncio.wait_for(future, timeout)
+            if msg.headers and msg.headers.get(STATUS_HDR
+                                               ) == NO_RESPONDERS_STATUS:
+                raise NoRespondersError
             return msg
         except asyncio.TimeoutError:
             await sub.unsubscribe()
