+++ nats/js/client.py
@@ -17,6 +17,7 @@ import json
 import time
 import nats.errors
 import nats.js.errors
+from nats.aio.subscription import Subscription
 from nats.js.manager import JetStreamManager
 from nats.js import api
 from typing import Any, Dict, List, Optional
@@ -139,8 +140,12 @@ class JetStream:
                 await js.subscribe('hello', cb=greetings)
 
                 # Durable Async Subscribe
-                # NOTE: Only one subscription can be bound to a durable name.
+                # NOTE: Only one subscription can be bound to a durable name. It also auto acks by default.
                 await js.subscribe('hello', cb=greetings, durable='foo')
+
+                # Durable Sync Subscribe
+                # NOTE: Sync subscribers do not auto ack.
+                await js.subscribe('hello', durable='bar')
 
                 # Queue Async Subscribe
                 # NOTE: Here 'workers' becomes deliver_group, durable name and queue name.
@@ -154,6 +159,7 @@ class JetStream:
             stream = await self._jsm.find_stream_name_by_subject(subject)
 
         deliver = None
+        consumer = None
 
         # If using a queue, that will be the consumer/durable name.
         if queue:
@@ -164,11 +170,25 @@ class JetStream:
             else:
                 durable = queue
 
-        try:
-            # TODO: Detect configuration drift with the consumer.
-            cinfo = await self._jsm.consumer_info(stream, durable)
-            config = cinfo.config
+        cinfo = None
+        consumerFound = False
+        shouldCreate = False
 
+        # Ephemeral subscribe always has to be auto created.
+        if not durable:
+            shouldCreate = True
+        else:
+            try:
+                # TODO: Detect configuration drift with any present durable consumer.
+                cinfo = await self._jsm.consumer_info(stream, durable)
+                config = cinfo.config
+                consumerFound = True
+                consumer = durable
+            except nats.js.errors.NotFoundError:
+                shouldCreate = True
+                consumerFound = False
+
+        if consumerFound:
             # At this point, we know the user wants push mode, and the JS consumer is
             # really push mode.
             if not config.deliver_group:
@@ -195,9 +215,8 @@ class JetStream:
                     raise nats.js.errors.Error(
                         f"cannot create a queue subscription {queue} for a consumer with a deliver group {config.deliver_group}"
                     )
-
-        except nats.js.errors.NotFoundError:
-            # If not found then attempt to create a consumer.
+        elif shouldCreate:
+            # Auto-create consumer if none found.
             if config is None:
                 # Defaults
                 config = api.ConsumerConfig(
@@ -219,8 +238,11 @@ class JetStream:
             # Auto created consumers use the filter subject.
             config.filter_subject = subject
 
-            await self._jsm.add_consumer(stream, config=config)
+            cinfo = await self._jsm.add_consumer(stream, config=config)
+            consumer = cinfo.name
 
+        # By default, async subscribers wrap the original callback and
+        # auto ack the messages as they are delivered.
         if cb and not manual_ack:
             ocb = cb
 
@@ -230,11 +252,10 @@ class JetStream:
 
             cb = new_cb
 
-        # TODO: Change into PushSubscription after refactoring types.
         sub = await self._nc.subscribe(
             config.deliver_subject, config.deliver_group, cb=cb
         )
-        return sub
+        return JetStream.PushSubscription(self, sub, stream, consumer)
 
     async def pull_subscribe(
         self,
@@ -287,6 +308,42 @@ class JetStream:
             return True
         else:
             return False
+
+    class PushSubscription(Subscription):
+        """
+        PushSubscription is a subscription that is delivered messages.
+        """
+        def __init__(self, js, sub, stream, consumer):
+            self._js = js
+            self._stream = stream
+            self._consumer = consumer
+
+            self._conn = sub._conn
+            self._id = sub._id
+            self._subject = sub._subject
+            self._queue = sub._queue
+            self._max_msgs = sub._max_msgs
+            self._received = sub._received
+            self._cb = sub._cb
+            self._future = sub._future
+            self._closed = sub._closed
+
+            # Per subscription message processor.
+            self._pending_msgs_limit = sub._pending_msgs_limit
+            self._pending_bytes_limit = sub._pending_bytes_limit
+            self._pending_queue = sub._pending_queue
+            self._pending_size = sub._pending_size
+            self._wait_for_msgs_task = sub._wait_for_msgs_task
+            self._message_iterator = sub._message_iterator
+
+        async def consumer_info(self):
+            """
+            consumer_info gets the current info of the consumer from this subscription.
+            """
+            info = await self._js._jsm.consumer_info(
+                self._stream, self._consumer
+            )
+            return info
 
     class PullSubscription:
         """
